move auth test to own file
