+++ test/peer-test.js
@@ -9,20 +9,9 @@ var testWsPort = 2315;
 
 var daemon;
 
-users = {};
-users['John Doe'] = {
-	password: 'foo',
-	auth: {
-		fetchGroups: ['public'],
-		setGroups: ['public'],
-		callGroups: ['public']
-	}
-};
-
 before(function () {
-	daemon = new jet.Daemon({
-		users: users
-	});
+	daemon = new jet.Daemon({});
+
 	daemon.listen({
 		tcpPort: testPort,
 		wsPort: testWsPort
@@ -144,46 +133,6 @@ describe('Jet module', function () {
 
 		it('should be an instance of EventEmitter', function () {
 			expect(peer).to.be.an.instanceof(EventEmitter);
-		});
-
-		it('can authenticate with a valid user/password', function (done) {
-			peer.authenticate('John Doe', 'foo', {
-				success: function (result) {
-					expect(result).to.deep.equal(users['John Doe'].auth);
-					done();
-				},
-				error: function (err) {
-					expect(err).to.be.an('undefined');
-				}
-			});
-		});
-
-		it('cannot authenticate with a valid user and invalid password', function (done) {
-			peer.authenticate('John Doe', 'foo2', {
-				success: function (result) {
-					expect(result).to.be.an('undefined');
-					done();
-				},
-				error: function (err) {
-					expect(err.message).to.equal('Invalid params');
-					expect(err.data).to.equal('invalid password');
-					done();
-				}
-			});
-		});
-
-		it('cannot authenticate with a invalid user and invalid password', function (done) {
-			peer.authenticate('John Doex', 'foo', {
-				success: function (result) {
-					expect(result).to.be.an('undefined');
-					done();
-				},
-				error: function (err) {
-					expect(err.message).to.equal('Invalid params');
-					expect(err.data).to.equal('invalid user');
-					done();
-				}
-			});
 		});
 
 		it('can add, fetch and set a state', function (done) {
